Simplify NamespaceRegistry
Don’t automatically prefix paths with `kernel.project_dir` anymore.

It was added as security precaution, but by defining a directory `/../../../test` you could still break out of the project directory. So this bad solution was removed.

Just like `open_basedir` in PHP, you should just solve this issue by using proper file and directory permissions instead.
Furthermore, as this is static (+ compiled) config, it is not an attack vector.

+ some CS