Allow older versions of Symfony/Mime to be used
Unfortunately we have to at least require v4.3 as there is no older version of it available. This should be still sufficient enough to support a wider range of apps as we're now still back into the 4.x area, as most 4.x restrictions will now allow to install this version of the AssetsBundle again.